+++ src/common/syncfilestatus.cpp
