+++ src/libsync/CMakeLists.txt
@@ -44,7 +44,6 @@ set(libsync_SRCS
     propagateremotemkdir.cpp
     syncengine.cpp
     syncfileitem.cpp
-    syncfilestatus.cpp
     syncfilestatustracker.cpp
     localdiscoverytracker.cpp
     syncresult.cpp
